requirements.txt
numpy
pandas

python-dotenv

vaderSentiment

langchain_anthropic
langchain_core

fastapi[standard]
